### 2024-09-20 #7248 @editorialbot
[REVIEW]: amisrsynthdata: A Python package for generating synthetic data for the Advanced Modular Incoherent Scatter Radars
**Submitting author:** @ljlamarche (<a href="http://orcid.org/0000-0001-7098-0524">Leslie Lamarche</a>)
**Repository:** https://github.com/amisr/amisrsynthdata
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.1.8
