### 2024-09-30 amisr/amisrsynthdata#13 @peijin94
Examples and use cases 
Although there are clear instructions and documents, the user can benefit a lot from some example use cases in the form of jupyternotebook.
