### 2024-09-30 amisr/amisrsynthdata#14 @peijin94
Benchmark and Performance consideration
There should be descriptions of how long a typical synthetic data run takes.

For performance consideration:
parsing txt file to number into numpy is much more time-consuming than direct loading .np or .npz file
Replacing beamcodes text file with h5 or npz can optimize the speed of `beams_from_beam_codes`
