### 2024-09-30 amisr/amisrsynthdata#15 @peijin94
pyproject.toml dependency package not complete
missing apexpy, gemini3d
