### 2024-09-30 amisr/amisrsynthdata#14 @peijin94
np.in1d has a deprecation error, better change to `isin`
