### 2024-09-30 amisr/amisrsynthdata#16 @peijin94
Summary plot in README and manuscript
It would be nice to have a 4 panel (N_e Ti Te Vlos) plot showing an example of the output plot of the synthesized data.
