### 2024-09-30 #7248 @peijin94
https://github.com/amisr/amisrsynthdata/issues/13
https://github.com/amisr/amisrsynthdata/issues/14
https://github.com/amisr/amisrsynthdata/issues/15
https://github.com/amisr/amisrsynthdata/issues/16
