### 2024-10-08 e0fbdda
Fix problem where excluding SUMMARY_PLOT section from config file crashes code.
