### 2024-10-09 0bfa129
Add apexpy as dependency.  Add pygemini as optional dependency and impliment nessisary code reorganization to avoid ModuleNotFound errors when NOT using GEMINI.
