+++ src/amisrsynthdata/state_functions/density.py
@@ -1,7 +1,8 @@
 import numpy as np
 import pymap3d as pm
 import datetime as dt
-from .utils import output_shape, gemini_helper
+#from .utils import output_shape, gemini_helper
+from .utils import output_shape
 
 
 class Density(object):
@@ -339,6 +340,10 @@ class Density(object):
         gemini_output_dir: string
             Path to directory of GEMINI output files
         """
+        # importing gemini_helper will trigger loading gemini3d, which
+        #   is only available if pygemini is installed as an optional
+        #   package requirement [gemini]
+        from .gemini_utils import gemini_helper
 
         gh = gemini_helper(self.gemini_output_dir, glat, glon, galt)
 
