### 2024-10-11 8766c95
Add some minor improvements to summary plots.
