### 2024-10-12 c4bd66d
Convert beamcode tables to single hdf5 file.
