### 2024-10-12 amisr/amisrsynthdata#14 @ljlamarche
The function np.in1d has been replaced with np.isin, as suggested. (37193986f4b24712f93c573c1098478c1937a149)

The beamcode *.txt files have been converted into a single h5 file. (c4bd66d38cee3b74519869ab819997e3c1ce989a)
