### 2024-10-12 45df529
Add a section in the documentation on benchmarking and describing approximately the time it take to run amisrsynthdata.
