### 2024-10-12 amisr/amisrsynthdata#14 @ljlamarche
The time required for synthetic data to be created depends on the ionospheric state that is chosen and the resolution of the radar mode.  A section on benchmarking has been added to the Usage section of the documentation. (45df529c81c3b91b6535e110674434cb8bbef74f)

|              | Simple | Complex |
|--------|--------|----------|
| No Plots | 0.77 s |  87.48 s |
| Plots      | 9.72 s | 103.24 s |
