### 2024-10-13 amisr/amisrsynthdata#13 @ljlamarche
A tutorial jupyter notebook has been added to the repository. (29ef2c08d4166f97215c4180f1568b288fb7570c)
