### 2024-10-14 25e5f8d
Add new figure to manuscript.
