### 2024-10-15 685dfa3
Update unit tests and correct some warnings.
