+++ src/amisrsynthdata/state_functions/density.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pymap3d as pm
 import datetime as dt
-#from .utils import output_shape, gemini_helper
 from .utils import output_shape
 
 
