### 2024-10-15 1d9ea54
Update parameter figure.
