### 2024-10-15 amisr/amisrsynthdata#16 @ljlamarche
A figure has been added showing all four parameters as altitude slices. (1d9ea542a0d1a6b05cc1b8e0e424fd23fe9aaefa)
