### 2024-10-20 #7248 @peijin94
Previously no benchmark and performance is provided, now it is completed, I have no more questions
