### 2024-10-21 amisr/amisrsynthdata#17 @rweigel
Installation instructions
I executed the two commands at

https://github.com/amisr/amisrsynthdata/tree/main?tab=readme-ov-file

```
pip install amisrsynthdata
amisrsynthdata config.yaml
```

and received an error related to apexpy not being installed. Perhaps warn the user that the second command will fail unless they do `pip install amisrsynthdata[apex]`?

At https://amisrsynthdata.readthedocs.io/en/latest/installation.html, the command `pin install` appears several times. Should that be `pip install`?
