### 2024-10-21 amisr/amisrsynthdata#17 @rweigel
The README gives

```
pip install amisrsynthdata[apex]
```

but I needed

```
pip install 'amisrsynthdata[apex]'
```
