### 2024-10-21 amisr/amisrsynthdata#17 @rweigel
```
git clone https://github.com/amisr/amisrsynthdata.git
cd amisrsynthdata
pip install -e .
```

gave

```
ERROR: File "setup.py" or "setup.cfg" not found. Directory cannot be installed in editable mode: /private/tmp/amisrsynthdata
(A "pyproject.toml" file was found, but editable mode currently requires a setuptools-based build.)
```
