### 2024-10-21 amisr/amisrsynthdata#15 @rweigel
When reviewing this, I expected to be able to copy/paste some commands on the command line and have something that just worked.

It appears that a prospective user must be aware of issues with dependencies or have a lot of time for trial and error in order to install this package and run an example.
