### 2024-11-05 6a34876
Remove the optional dependeny pygemini because pypi will not allow github dependencies.
