+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "amisrsynthdata"
-version = "1.1.3"
+version = "1.1.4"
 authors = [
   { name="L. Lamarche", email="leslie.lamarche@sri.com" },
 ]
