### 2024-11-06 amisr/amisrsynthdata#17 @ljlamarche
Hi @rweigel , thank you for your comments!  I've published a new release of this package to pypi that fixes many of these installation issues.  The one caveat is the apexpy dependency has some version incompatibilities (https://github.com/aburrell/apexpy/issues/134, https://github.com/aburrell/apexpy/issues/135) that aren't currently properly accounted for.  Sometime next week, I think we'll do a new minor release of apexpy with required versions pinned, which should address the issue of amirsynthdata not installing easily with pip.  For now, if you'd like to try reinstalling, please do the following in python 3.11 or lower:

```
pip install numpy==1.26.4
pip install 'amisrsynthdata[plots]'
```

I will post here when the new version of apexpy has been published and you can try installing with the normal instructions.
