### 2024-11-11 amisr/amisrsynthdata#17 @rweigel
I was able to install, but wanted to execute something like

```
amisrsynthdata config.yaml
```

but it seems I need to use the `pip install -e ...` example to access this file. After doing so, I was not able to find `config.yaml`, but did find `example_synth_config.yaml`. I recommend clarification in the documentation. Perhaps set this file as the default when `amisrsynthdata` is executed and inform the user that it is reading the default and print the location of the default. This will also make the user not need to do the `pip install -e` - they can execute `amisrsynthdata`, see the path of the default, copy it locally and edit it, and then run again with the modified file.

As a reviewer (and user), I would prefer to not spend most of my time dealing with installation and packaging issues before getting something working. I recommend adding a sequence of commands in the documentation that will 'just work' and for which statements such as 'This assumes you already have [apexpy](https://apexpy.readthedocs.io/en/latest/) installed' are not needed. (See the "don't make me think" principle.)

One minor recommendation. `amisrsynthdata example_synth_config.yaml` gives the feedback
```
/opt/miniconda3/envs/python3.11.9/lib/python3.11/site-packages/cartopy/io/__init__.py:241: DownloadWarning: Downloading: https://naturalearth.s3.amazonaws.com/10m_physical/ne_10m_coastline.zip
  warnings.warn(f'Downloading: {url}', DownloadWarning)
```

which is useful, but I am not sure why this is a warning. Also, after the program completed, no feedback was given and my first guess was it failed. However, several files were created in the repo. I suggest giving more feedback to the user.

Also, I sure if there is a way of handling this, but I tried the commands without checking my Python version (which was 3.12) and after about 30 seconds received an error. Is there a way to abort the installation early if there is a python version incompatibility?
