+++ docs/source/usage.rst
@@ -13,7 +13,7 @@ To run the AMISR synthetic data generator, simply call ``amisrsynthdata`` from t
   amisrsynthdata config.yaml
 
 
-See :ref:`Configuration File` section of the documentation for details on the fields that should be in this file.
+See :ref:`Configuration File` section of the documentation for details on the fields that should be in this file.  An example configuration file can be downloaded from the `root of the GitHub repository <https://github.com/amisr/amisrsynthdata>`_.
 
 Python
 ------
