+++ docs/requirements.txt
@@ -1,1 +1,2 @@
+sphinx>=1.3
 sphinx_rtd_theme
