+++ .readthedocs.yaml
@@ -4,12 +4,10 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.9"
+    python: "3.11"
 
 python:
   install:
-    - method: pip
-      path: .
     - requirements: docs/requirements.txt
 
 # Build documentation in the docs/ directory with Sphinx
