+++ .readthedocs.yaml
@@ -12,4 +12,4 @@ python:
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-   configuration: docs/conf.py
+   configuration: docs/source/conf.py
