+++ src/amisrsynthdata/syntheticdata.py
@@ -737,12 +737,17 @@ def main():
     parser.add_argument(
         'synth_config_file',
         help='Configuration file for synthetic data set.')
+    parser.add_argument('-v', '--verbose', action='store_true')
     args = vars(parser.parse_args())
 
     with open(args['synth_config_file'], 'r') as cf:
         config = yaml.safe_load(cf)
 
+    if args['verbose']: print('Initializing synthetic data module...')
     sd = SyntheticData(config)
+    if args['verbose']: print('Generating output file...')
     sd.create_hdf5_output(config['GENERAL']['output_filename'])
     if 'SUMMARY_PLOT' in config:
+        if args['verbose']: print('Generating summary plots...')
         sd.create_summary_plots(**config['SUMMARY_PLOT'])
+    if args['verbose']: print('amisrsynthdata has finished!')
