+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "amisrsynthdata"
-version = "1.1.7"
+version = "1.1.8"
 authors = [
   { name="L. Lamarche", email="leslie.lamarche@sri.com" },
 ]
