### 2024-11-17 amisr/amisrsynthdata#17 @ljlamarche
Thanks you for the feedback on the installation issues!  I've updated the documentation to provide clearer starting instructions for new users.  I've [published a new version (v1.1.8) to pypi](https://pypi.org/project/amisrsynthdata/), which should link to the ["stable" (default) documentation on ReadTheDocs](https://amisrsynthdata.readthedocs.io/en/stable/).  Additionally, a new version of apexpy has been published with versions pinned so pip installation should "just work" now.  Can you please try the [new instructions](https://amisrsynthdata.readthedocs.io/en/stable/#example-usage) and let me know if you have issues?

I'm hesitant include the example configuration file in the package because it doesn't make a lot of sense for this code and it's application to have a "default".  There isn't really a single background ionospheric state or radar mode that is "typical", and I worry that introducing users to the package with a built-in configuration file but requiring they provide a different one as soon as they start doing anything "real" with the package would be confusing.  I have modified the documentation to include more specific instruction of how to download and use the example configuration file.

The warning you identified comes from cartopy - it looks like it's just informing you that it is downloading coastline data for the plots.  I believe this is standard for cartopy and I'm not sure I should suppress the warning in amisrsynthdata.

Unfortunately, I'm not sure there is a way to check the python version before building the package and save yourself the 30 seconds of failed installation time... This seems to be a feature of pip and I can't find any flag or setting to check the python version first.  I have pinned python<3.12 in both amisrsynthdata and apexpy so it's evident on the pypi page at least.  I know we're hoping to fix the apexpy compatibility issue sometime before the end of the year.
