### 2024-11-18 #7248 @rweigel
@mbobra - I have been active at https://github.com/amisr/amisrsynthdata/issues/17 over the past month.

It is close to satisfying the checklist requirements. I expect to look at this tomorrow.
