### 2024-11-19 #7248 @rweigel
I have completed my review.

The manuscript is well-written and clear, and the software provides a significant algorithm that is justified in the manuscript. I expect this software to be useful for validating and learning about ASMIR measurements.
