### 2024-11-19 amisr/amisrsynthdata#17 @rweigel
I have finished [my checklist](https://github.com/openjournals/joss-reviews/issues/7248#issuecomment-2485125013).

The following notes are for the benefit of the author and do not need a reply or action.

See https://stackoverflow.com/questions/19534896/enforcing-python-version-in-setup-py regarding Python version checking. I put version checks in my code when I know something will fail because I've found many users conclude "code does not work" and give up and move on when a pip install does not work. I don't regard this as mandatory, however. Probably this should be handled in ApexPy.

https://amisrsynthdata.readthedocs.io/en/stable/installation.html# has `pip install amisrsynthdata[plots]`, which does not work for me without quotes.

I was able to install using `pip install 'amisrsynthdata[plots]'`

Other optional edits to consider:
* python -> Python in manuscript.
* I usually use `curl -O` instead of `wget` in my examples because I've found `curl` always installed already on Linux, Mac, and (more recently) Windows, whereas `wget` usually requires an install.
* At https://amisrsynthdata.readthedocs.io/en/stable/#example-usage, docs/synthdata_summary_ne.png docs/synthdata_gemini_plot.png show a broken image.
* Add links where "SRI data format" is mentioned (at least twice) because a Google search on "SRI data format" does not turn up anything relevant on the first page.
* Link to your suggestion for ApexPy install issues in the example usage section
* Link to the JOSS paper at https://amisrsynthdata.readthedocs.io/en/stable/# (When I am writing a paper, I always appreciate not having to spend a lot of time figuring out how to cite software and data.)
* Remove empty page (or was there a generation error)? https://amisrsynthdata.readthedocs.io/en/stable/reference.html
* Is https://amisrsynthdata.readthedocs.io/en/stable/usage.html#syntheticdata-class-example expected to work with the example config? I get  `fov_glat = sd.Geomag['Latitude'] AttributeError: 'SyntheticData' object has no attribute 'Geomag'`
* I find it helpful if a link to the GitHub repo is prominent in documentation, for example at https://amisrsynthdata.readthedocs.io/en/stable/index.html. I typically read the docs and scan the source before attempting to use something, so cross-links are appreciated.
* Example config and test config are similar. It may help reduce future maintenance if only test file used as example.
* Link to tutorial notebook in https://amisrsynthdata.readthedocs.io/en/stable/index.html#example-usage
* About 1/2 of page at https://github.com/amisr/amisrsynthdata/blob/main/tutorial.ipynb is warnings.

----

I doubt this is worth fixing, but on an older intel mac 10.15 with Python 3.10 I got the following (I tend to not upgrade OS on systems with many difficult-to install packages, such as ApexPy, because an upgrade means days of figuring out how to get the packages working again).

```
gfortran --version
GNU Fortran (Homebrew GCC 9.3.0_1) 9.3.0
Copyright (C) 2019 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```

```
pip install 'amisrsynthdata[plots]'
Collecting amisrsynthdata[plots]
  Downloading amisrsynthdata-1.1.8-py3-none-any.whl (107 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 107.1/107.1 kB 3.8 MB/s eta 0:00:00
Collecting pymap3d
  Downloading pymap3d-3.1.0-py3-none-any.whl (60 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 60.6/60.6 kB 3.6 MB/s eta 0:00:00
Collecting apexpy
  Downloading apexpy-2.0.2.tar.gz (331 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 331.4/331.4 kB 10.7 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Preparing metadata (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [16 lines of output]
      + meson setup /private/var/folders/pr/6xm0lrgx50g4276btj3p_dh00000gn/T/pip-install-l8vlg7vw/apexpy_7916eb0a5b27429f9b30aee37f30ec3b /private/var/folders/pr/6xm0lrgx50g4276btj3p_dh00000gn/T/pip-install-l8vlg7vw/apexpy_7916eb0a5b27429f9b30aee37f30ec3b/.mesonpy-qpwie1xo -Dbuildtype=release -Db_ndebug=if-release -Db_vscrt=md --native-file=/private/var/folders/pr/6xm0lrgx50g4276btj3p_dh00000gn/T/pip-install-l8vlg7vw/apexpy_7916eb0a5b27429f9b30aee37f30ec3b/.mesonpy-qpwie1xo/meson-python-native-file.ini
      The Meson build system
      Version: 1.6.0
      Source dir: /private/var/folders/pr/6xm0lrgx50g4276btj3p_dh00000gn/T/pip-install-l8vlg7vw/apexpy_7916eb0a5b27429f9b30aee37f30ec3b
      Build dir: /private/var/folders/pr/6xm0lrgx50g4276btj3p_dh00000gn/T/pip-install-l8vlg7vw/apexpy_7916eb0a5b27429f9b30aee37f30ec3b/.mesonpy-qpwie1xo
      Build type: native build
      Project name: apexpy
      Project version: 2.0.2
      C compiler for the host machine: cc (clang 12.0.0 "Apple clang version 12.0.0 (clang-1200.0.32.29)")
      C linker for the host machine: cc ld64 609.8
      Host machine cpu family: x86_64
      Host machine cpu: x86_64

      ../meson.build:17:0: ERROR: Compiler gfortran cannot compile programs.

      A full log can be found at /private/var/folders/pr/6xm0lrgx50g4276btj3p_dh00000gn/T/pip-install-l8vlg7vw/apexpy_7916eb0a5b27429f9b30aee37f30ec3b/.mesonpy-qpwie1xo/meson-logs/meson-log.txt
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```
