### 2024-11-19 amisr/amisrsynthdata#17 @ljlamarche
@rweigel , thank you for your thorough feedback!  I will work to fix these point by point.  The documentation and linking issues can definitely be corrected, I appreciate you identifying them.

The dependency on the finicky apexpy is an unfortunate necessity.  I really need a magnetic coordinate package to handle the plasma velocity correctly, and apexpy has certain capabilities that tend to be lacking from the alternative options (most of which are equally if not more fraught with installation issues).  I will log the issues you have had and we will try to fix them in the next release of apexpy.
