### 2024-11-20 #7248 @mbobra
Thank you both so very much for your thorough reviews, @rweigel and @peijin94! I know you put a lot of effort into this. Your volunteer time keeps JOSS running and we sincerely appreciate your time and effort.
