### 2024-11-20 #7248 @mbobra
Congratulations, @ljlamarche -- we're almost ready to publish the paper! Could you please make sure that the title on the [Zenodo deposit](https://zenodo.org/records/14174414) matches the title on the JOSS paper? Then I can move forward with accepting the paper.
