### 2024-11-22 #7248 @ljlamarche
Thank you @mbobra!  I'm a bit confused.  I believe that zenodo deposit is the automatic archiving of the source code.  Would you like me to make a separate zenodo repo for the paper itself?
