### 2024-11-22 6fe0cba
Modify .readthedocs.yaml so the package will be installed for autodocs.
