+++ .readthedocs.yaml
@@ -8,6 +8,8 @@ build:
 
 python:
   install:
+    - method: pip
+      path: .
     - requirements: docs/requirements.txt
 
 # Build documentation in the docs/ directory with Sphinx
