+++ .readthedocs.yaml
@@ -8,8 +8,6 @@ build:
 
 python:
   install:
-    - method: pip
-      path: .
     - requirements: docs/requirements.txt
 
 # Build documentation in the docs/ directory with Sphinx
