### 2024-11-22 #7248 @mbobra
@ljlamarche Ah sorry, not at all. Right now, the title of your Zenodo deposit is _amisr/amisrsynthdata: v1.1.8_. I'm requesting you change the title, by manually editing [this page](https://zenodo.org/records/14174414), to _amisrsynthdata: A Python package for generating synthetic data for the Advanced Modular Incoherent Scatter Radars_. This way the title of the JOSS paper and Zenodo deposit match exactly, and people can easily tell which version of the software matches up with the description of the package in the JOSS paper.
