### 2024-11-22 b4c6242
Add setup.py to handle case of not installing apexpy for RTDs.
