### 2024-11-26 #7248 @ljlamarche
@mbobra , the proof looks good to me!  Do I need to enter the accept command or you?
