### 2024-11-26 #7248 @warrickball
Hi @ljlamarche — I'm the assistant editor in chief for this subject area and am working on some editorial changes. One thing that needs addressing is that the software licence in the Zenodo archive (currently CC-BY-4.0) should be changed to match the software repo (GPLv3).

I also noticed you've cited a particular version of h5py via [this Zenodo archive](https://zenodo.org/records/6575970). If you wish, you could use the [DOI for all versions of that repository](https://doi.org/10.5281/zenodo.594310), which will always resolve to the latest version on Zenodo. I won't change this but you may want to.

You should see a PR with my editorial tweaks in the next few minutes.
