### 2024-11-26 3bd27be
Editorial tweaks for publication in JOSS
