+++ manuscript/paper.bib
@@ -155,7 +155,7 @@
 
 @manual{cartopy,
 author = {{Met Office}},
-title = {Cartopy: a cartographic python library with a Matplotlib interface},
+title = {{Cartopy}: a cartographic {Python} library with a {Matplotlib} interface},
 year = {2010 - 2015},
 address = {Exeter, Devon },
 url = {https://scitools.org.uk/cartopy}
