+++ manuscript/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 Advanced Modular Incoherent Scatter Radars (AMISR) are a class of ground-based radars that are used to study plasma dynamics in the ionosphere [@Kelly2009].  Ionospheric plasma is highly structured and varies on a wide range of scales, both spatially and temporally. There are large and often abrupt variations in the plasma state parameters depending on geomagnetic conditions and local or global events. AMISR remotely measures the electron density, $N_e$, ion temperature, $T_i$, electron temperature, $T_e$, and line-of-sight component of the drift velocity, $V_{los}$, of the plasma in the ionosphere.  These ionospheric state parameters are measured in profiles over different look directions over time.
 
-AMISR utilizes phased-array technology to steer the radar beam electronically, which allows it to change look directions rapidly with a sub-second timescale.  This allows the radar to both effectively collect data from different look directions nearly simultaneously and have a great degree of flexibility in terms of beam positions and observational mode designs.  Most standard modes use between 4 and 50 beams, roughly evenly spread across the radar field-of-view.  Some specialized modes use tight clusters of beams or other unusual configurations to observe specific phenomena. This flexibility, as well as the radars' locations in regions that are key for magnetosphere-ionosphere coupling, make them a heavily utilized and invaluable resource for space physics research.
+AMISR utilizes phased-array technology to steer the radar beam electronically, which allows it to change look directions rapidly with a sub-second timescale.  This allows the radar to both effectively collect data from different look directions nearly simultaneously and have a great degree of flexibility in terms of beam positions and observational mode designs.  Most standard modes use between 4 and 50 beams, roughly evenly spread across the radar field-of-view.  Some specialized modes use tight clusters of beams, or other unusual configurations to observe specific phenomena. This flexibility, as well as the radars' locations in regions that are key for magnetosphere-ionosphere coupling, make them a heavily utilized and invaluable resource for space physics research.
 
 ![Left: A 3D visualization of beam positions with synthetic density measurements from a commonly run 11 beam AMISR mode. Right: All possible beam positions within the AMISR field-of-view.](amisr_fov.png)
 
