### 2024-11-26 #7248 @warrickball
I've opened [a PR](https://github.com/amisr/amisrsynthdata/pull/21) with some tweaks. With those implemented and the archive licence updated, I'll do another test run of our publication pipeline.
