### 2024-11-26 #7248 @warrickball
As an aside, I noticed that your CI reports failures on one set of tests. To me, this simply looks like small variation from differences in floating point numbers. If you call [`numpy.testing.assert_allclose`](https://numpy.org/doc/2.1/reference/generated/numpy.testing.assert_allclose.html) with a small absolute tolerance (`atol`, 1e-12 should do) then I think your tests will pass in Python 3.7–3.11.
