### 2024-12-02 0f330e0
Add setup.py in an attempt to fix RTD autodocs issue (NOT WORKING YET).
