+++ manuscript/paper.bib
@@ -127,13 +127,11 @@
                   Daniel Hay Guest and
                   Yu Feng and
                   Mark Kittisopikul},
-  title        = {h5py/h5py: 3.7.0},
-  month        = may,
-  year         = 2022,
+  title        = {h5py},
+  year         = 2023,
   publisher    = {Zenodo},
-  version      = {3.7.0},
-  doi          = {10.5281/zenodo.6575970},
-  url          = {https://doi.org/10.5281/zenodo.6575970}
+  doi          = {10.5281/zenodo.594310},
+  url          = {https://doi.org/10.5281/zenodo.594310}
 }
 
 @article{pymap3d, doi = {10.21105/joss.00580}, url = {https://doi.org/10.21105/joss.00580}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {23}, pages = {580}, author = {Michael Hirsch}, title = {PyMap3D: 3-D coordinate conversions for terrestrial and geospace environments}, journal = {Journal of Open Source Software} } 
