### 2024-12-03 53db09e
New attempt at fixing autodocs by specifying dependencies in setup.py so apexpy can be removed dynamically in the RTD environment.
