### 2024-12-03 ac5c526
Remove python 3.12 from unit test matrix - fails due to known apexpy installation problem.
