### 2024-12-03 207747d
Fix linter issues in setup.py.
