### 2024-12-03 #7248 @ljlamarche
@warrickball, thank you for all your recommendations!  I've fixed the license in the zenodo archive and accepted your PR.  I also changed the h5py citation and implemented your corrections to the unit test (they should all pass now).  Please let me know if there's anything else!
