### 2024-12-05 #7248 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
