### 2024-12-05 #7248 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
